#19 New part traits to help with insights
